fix: split wp ai client cache classes
